+++ src/main/java/net/frozenblock/lib/config/clothconfig/mixin/client/DynamicEntryListWidgetEntryMixin.java
@@ -25,9 +25,9 @@ import me.shedaniel.clothconfig2.gui.widget.DynamicEntryListWidget;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.frozenblock.lib.FrozenLogUtils;
-import net.frozenblock.lib.config.api.annotation.FieldIdentifier;
 import net.frozenblock.lib.config.api.instance.Config;
 import net.frozenblock.lib.config.api.instance.ConfigModification;
+import net.frozenblock.lib.config.api.sync.annotation.EntrySyncData;
 import net.frozenblock.lib.config.clothconfig.impl.DisableableWidgetInterface;
 import net.frozenblock.lib.config.impl.network.ConfigSyncModification;
 import org.jetbrains.annotations.NotNull;
@@ -42,10 +42,12 @@ public abstract class DynamicEntryListWidgetEntryMixin implements DisableableWid
 	private ConfigModification.EntryPermissionType frozenLib$entryPermissionType = ConfigModification.EntryPermissionType.CAN_MODIFY;
 
 	@Override
-    public void frozenLib$addSyncData(@NotNull Class<?> clazz, String identifier, Config<?> configInstance) {
+    public void frozenLib$addSyncData(@NotNull Class<?> clazz, @NotNull String identifier, Config<?> configInstance) {
+		if (identifier.equals("")) new Exception("Cannot process sync value with empty identifier!").printStackTrace();
 		Field field = null;
 		for (Field fieldToCheck : clazz.getDeclaredFields()) {
-			if (fieldToCheck.isAnnotationPresent(FieldIdentifier.class) && fieldToCheck.getAnnotation(FieldIdentifier.class).identifier().equals(identifier)) {
+			EntrySyncData entrySyncData = field.getAnnotation(EntrySyncData.class);
+			if (entrySyncData!= null && !entrySyncData.value().equals("") && entrySyncData.value().equals(identifier)) {
 				if (field != null) FrozenLogUtils.logError("Multiple fields in " + clazz.getName() + " contain identifier " + identifier + "!", true, null);
 				field = fieldToCheck;
 			}