+++ CHANGELOG.md
@@ -7,3 +7,5 @@ Put changelog here:
 -----------------
 - Updated embedded Jankson
   - Adds `SaveToggle` annotation
+- Moved all field annotations for config syncing to `EntrySyncData` for easier use
+  - Fixed a small bug with `Locked When Synced` fields syncing instead of remaining stagnant